third_party/blink/web_tests/external/wpt/sanitizer-api/sanitizer-custom-elements-is.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] The is= attribute is removed by the default sanitizer config.
  assert_equals: expected "<div>hello</div>" but got "<div is=\\"foo-bar\\">hello</div>"
[FAIL] The is= attribute is removed by the global removeAttributes
  assert_equals: expected "<div>hello</div>" but got "<div is=\\"foo-bar\\">hello</div>"
[FAIL] The is= attribute is removed by the local removeAttributes
  assert_equals: expected "<div>hello</div>" but got "<div is=\\"foo-bar\\">hello</div>"
[FAIL] The is= attribute is removed by the local missing attributes
  assert_equals: expected "<div>hello</div>" but got "<div is=\\"foo-bar\\">hello</div>"
Harness: the test ran to completion.
